// src/app/shared/enums/modalidade.ts
export enum Modalidade{
  BEACHTENNIS = 'BEACH TENNIS',
  DANCA = 'DANCA',
  FUNCIONAL = 'FUNCIONAL',
  FUTEBOL = 'FUTEBOL',
  FUTSAL = 'FUTSAL',
  FUTVOLEI = 'FUTVOLEI',
  GINASTICA = 'GINASTICA',
  JIUJITSU = 'JIU JITSU',
  MUAITHAI = 'MUAY THAI',
  NATAÇÃO = 'NATAÇÃO',
  PILATES = 'PILATES',
  TENIS = 'TENIS',
  VOLEI = 'VOLEI',
}